+++ src/values/RecorderValues.js
@@ -15,16 +15,40 @@ export const MEDIA_RECORDER_OPTIONS = {
   mimeType: MEDIA_RECORDER_SUPPORTED_MIME_TYPE,
 };
 
-/** @type {MediaStreamConstraints} */
-export const MEDIA_STREAM_CONSTRAINTS = {
-  video: {
-    facingMode: 'environment',
-    width: { ideal: 7680 },
-    height: { ideal: 4320 },
-    aspectRatio: { ideal: 16 / 9 },
+/** @type {Array<MediaStreamConstraints>} */
+export const MEDIA_STREAM_CONSTRAINTS = [
+  {
+    video: {
+      facingMode: 'environment',
+      width: { ideal: 7680 },
+      height: { ideal: 4320 },
+      aspectRatio: { exact: 16 / 9 },
+    },
+    audio: true,
   },
-  audio: true,
-};
+  {
+    video: {
+      facingMode: 'environment',
+      width: { ideal: 3840 },
+      height: { ideal: 2160 },
+      aspectRatio: { exact: 16 / 9 },
+    },
+    audio: true,
+  },
+  {
+    video: {
+      facingMode: 'environment',
+      width: { ideal: 1920 },
+      height: { ideal: 1080 },
+      aspectRatio: { ideal: 16 / 9 },
+    },
+    audio: true,
+  },
+  {
+    video: true,
+    audio: true,
+  },
+];
 
 /** @type {BlobPropertyBag} */
 export const MEDIA_BLOB_OPTIONS = {